core: Make Font ascent and descent i16, as it can technically be negative

+++ core/src/font.rs
@@ -143,11 +143,11 @@ struct FontData {
 
     /// The distance from the top of each glyph to the baseline of the font, in
     /// EM-square coordinates.
-    ascent: u16,
+    ascent: i16,
 
     /// The distance from the baseline of the font to the bottom of each glyph,
     /// in EM-square coordinates.
-    descent: u16,
+    descent: i16,
 
     /// The distance between the bottom of any one glyph and the top of
     /// another, in EM-square coordinates.
@@ -169,7 +169,7 @@ impl<'gc> Font<'gc> {
 
         let descriptor = FontDescriptor::from_swf_tag(&tag, encoding);
         let (ascent, descent, leading) = if let Some(layout) = &tag.layout {
-            (layout.ascent, layout.descent, layout.leading)
+            (layout.ascent as i16, layout.descent as i16, layout.leading)
         } else {
             (0, 0, 0)
         };
